Extract s3 utils
